Add files via upload
Spring framework provides an option to autowire the beans. In that case you don't need to explicitly wire the bean properties (using ref attribute) but Spring will do it automatically by using the "autowire" attribute.

In Spring you can autowire dependencies using XML configuration or use the annotations to autowire the dependencies. This post talks about autowiring in Spring using XML configuration.

Table of contents
Autowiring modes in Spring
Autowiring in Spring - XML configuration
Autowiring in Spring using byName
Autowiring in Spring using byType
Solving UnsatisfiedDependencyException with Spring autowiring
Autowiring in Spring by constructor
default-autowire attribute

Autowiring modes in Spring
Spring provides four kind of autowiring modes (updated as per Spring 4.x where autodetect option has been removed).

no- The traditional Spring default. No automagical wiring. Bean references must be defined in the XML file via the <ref/> element (or "ref" attribute). autowire=”default” is same as autowire=”no”

byName- Autowiring by name means properties of the autowired bean will be wired by searching for bean with same name/id in the configuration file. If such a bean is found it is injected, properties for which no matching bean is found will remain unwired which may later result in error.

byType- Autowiring by type means properties of the autowired bean will be wired by searching for bean whose type is compatible with the property's type. If such a bean is found it is injected, properties for which no matching bean is found will remain unwired which may later result in error.

constructor- In this case wiring will be done using the constructor of the autowired bean by looking for the bean whose type is compatible with the constructor argument.

autodetect- In case of autodetect auto wiring using constructor will be tried first, if default constructor only is there and autowiring using constructor can't be done then autowiring byType is used. Note that this option was available till Spring 3.x (deprecated) or older versions. In Spring 4.x it is removed and this option is not available anymore. In this post autodetect example is not there.
Autowiring in Spring - XML configuration
First thing to note here is that by default Spring doesn't autowire and you have to provide explicit wiring using ref attribute for the properties or for constructor-args which are to be dependency injected.

Refer Setter-based dependency injection in Spring to see an example of setter based dependency injection.
and
Refer Constructor-based dependency injection in Spring to see an example of constructor based dependency injection.